Betting Stage of Game Complete

+++ BlackjackServer/src/com/amarasapps/BlackjackServer.java
@@ -25,10 +25,10 @@ public class BlackjackServer {
     private static String SERVER_ADDRESS;               //IP Address of Server
 
     private static final int PLAYERS_PER_TABLE = 2;     //Number of Connected Clients
-    private static final int MINIMUM_BET = 100;         //Minimum Allowed Bet
+    private static final double MINIMUM_BET = 100.00;         //Minimum Allowed Bet
     private static final int DECKS_USED = 8;            //Decks Kept in the Shoe
     private static final int CARDS_BEFORE_SHUFFLE = 80; //Cards remaining in the deck before a re-shuffle
-    private static final int STARTING_MONEY = 500;      //Money that Each Player Would STart With
+    private static final double STARTING_MONEY = 500.00;      //Money that Each Player Would STart With
 
     public static void main(String[] args) {
